Replaced `int` types with `long`

+++ src/regular_polygon2d.cpp
@@ -15,12 +15,12 @@ void RegularPolygon2D::_bind_methods() {
     ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "radius"), "set_radius", "get_radius");
 }
 
-void RegularPolygon2D::set_nsides(int nsides_p) {
+void RegularPolygon2D::set_nsides(long nsides_p) {
     nsides = nsides_p;
     update_polygon();
 }
 
-int RegularPolygon2D::get_nsides() {
+long RegularPolygon2D::get_nsides() {
     return nsides;
 }
